# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "api/pressings/facture": {
            "get": {
                "tags": [
                    "Pressing"
                ],
                "summary": "Get a list of Factures",
                "description": "Returns a list of Facture objects based on the authenticated user's role",
                "operationId": "getFactures",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Invalid user role"
                    }
                }
            }
        },
        "api/pressings/facture/{id}": {
            "get": {
                "tags": [
                    "Pressing"
                ],
                "summary": "Get a Facture by ID",
                "description": "Returns a Facture object based on the given ID",
                "operationId": "getFactureById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Facture to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Facture not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Pressing"
                ],
                "summary": "Update a Facture by ID",
                "description": "Updates a Facture object based on the given ID",
                "operationId": "updateFactureById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Facture to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "JSON object containing data to update Facture",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request data"
                    },
                    "404": {
                        "description": "Facture not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pressing"
                ],
                "summary": "Delete a facture by ID",
                "description": "Delete a facture by ID. Requires admin role, or client_id or pressing_id must match authenticated user's ID.",
                "operationId": "deleteFacture",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the facture to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Facture deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Facture deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "You are not authorized to delete this facture",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "You are not authorized to delete this facture"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Facture not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Facture not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    }
}